+++ HISTORY.rst
@@ -9,5 +9,5 @@ History
 0.1.0.dev0
 ---------------------
 
-* Initial version - code from Galaxy's test framework with changes 
+* Initial version - code from Galaxy's test framework with changes
   based on planemo testing.